Refactor: Consolidate default options in one place.
Previously, default options were scattered and checked throughout
the code. It was hard to see what all the options were, and there
were extra conditional checks because the code couldn't depend an option
already being set or not.

Now the all defaults are declared in one place, and some related
conditional code can be removed, since the code can now expect
the defaults to be set in advance.

(the phantomFlag support has not been introduced yet, but this lays the
foundation for it).